Use ID_INPUT_KEY udev variable to detect "keyboard"
There are some keyboard-like devices (like multimedia keys, apparently
visible as separate device on some hardware), that still needs to
produce keyboard events (and the user want to allow them), but do not
have all the "normal" keys, so udev do not consider it a "keyboard". For
such devices it sets ID_INPUT_KEY type, so lets use it. Note that
ID_INPUT_KEYBOARD is set in addition to ID_INPUT_KEY, so we don't need
to check both.

Fixes QubesOS/qubes-issues#2781