tests: drop right and middle buttons from emulated touchscreen
Emulating tablet device requires specified precision - which is not
supported by python-uinput. Otherwise Xorg refuse to use it as tablet.
On the other hand, touchscreen does not require it, but also udevd in
Fedora >= 28 require that touchscreen has no mouse buttons, otherwise it
tags it as a mouse.

To work with that, modify tablet events to include only BTN_TOUCH.