Automatically detect dom0
qubes-input-trigger can be called with --dom0 option to adjust list of
devices it acts on. This option is correctly used on GUI VM startup, but
not when called from /etc/xdg/autostart/qubes-input-trigger.desktop
(which is the same in dom0 and VM). Fix this by automatically defaulting
to --dom0 if qubes-input-trigger is called in dom0.

Specifically, this will exclude virtual input devices that are passed
through from sys-usb already.