Avoid forwarding dom0 devices back to dom0
The input proxy tries to forward dom0 devices to dedicated sys-gui-gpu,
but if it isn't set (or started), they should remain in dom0 (that
is not an error). Avoid setting service state to failed in this case
(which would happen if qrexec-client gets called this way - as it can't
run dom0->dom0 services).