input-proxy-sender: Mask LED events
If LED events are not masked, the kernel feeds the LED events back into
the input-proxy-sender, even if it were the input-proxy-sender that
injected the LED events into the kernel in the first place.

This is because input-proxy-sender is a "client" of the "evdev" driver
and has an active "grab".

Quoting from Linux v6.6.39's "drivers/input/evdev.c":

```
288  /*
289   * Pass incoming events to all connected clients.
290   */
291  static void evdev_events(struct input_handle *handle,
292                           const struct input_value *vals, unsigned int count)
293  {
294          struct evdev *evdev = handle->private;
295          struct evdev_client *client;
296          ktime_t *ev_time = input_get_timestamp(handle->dev);
297
298          rcu_read_lock();
299
300          client = rcu_dereference(evdev->grab);
301
302          if (client)
303                  evdev_pass_values(client, vals, count, ev_time);
304          else
305                  list_for_each_entry_rcu(client, &evdev->client_list, node)
306                          evdev_pass_values(client, vals, count, ev_time);
307
308          rcu_read_unlock();
309  }
```

To prevent having the evdev driver reinject the LED events into the
input-proxy-sender and hence to avoid the avalanche of LED events that
eventually causes CPU-bound kworker thread(s) in "sys-usb", filter out
the LED events in input-proxy-sender, via an ioctl.

There is a possibility that other events need to be filtered out as
well, but we opt for filtering out the LED events only to fix the issue
reported in QubesOS/qubes-issues#5823 with a minimal change.

The issue has been reproduced by the author of this commit with a
Logitech Bolt USB receiver and/or with a wired external USB keyboard.
This commit has been used on the author's computer for a few months
without negative effects. Without this commit, the CPU-bound kworker
thread issue in sys-usb can be reproduced by rapidly pressing the
Caps-Lock key several times OR by repeatedly restarting the
input-proxy-sender daemons several times in sys-usb via systemctl, with
an LED event capable USB keyboard or other similar USB device connected.

Fixes: QubesOS/qubes-issues#5823
Signed-off-by: M. Vefa Bicakci <m.v.b@runbox.com>