Add device info to qrexec call argument
This allows setting policy for individual devices, not only device
types. It relies on USB qube (or wherever the device is connected)
giving accurate device info, but since it's going to be granted at least
partial control it needs to be trusted to some degree anyway.

Build the device info based on type/vendor/product/revision (the PRODUCT
property) and physical location (USB controller + port in case of USB
device). The latter can be disabled, to build a policy that allows
given device in any port.

Fixes QubesOS/qubes-issues#3604